Lexicon API

Lexicons

CJKFrequencies.Lexicon — Type

Lexicon()
Lexicon(io_or_filename)
Lexicon(words)

Construct a lexicon. It can be empty (no parameters) or created from some IO-like object or a sequence/iterable of words.

A lexicon is a list of (known) words, each of which can be tagged with various tags (e.g. indicating how it is known, etc.).

CJKFrequencies.tagged_with — Function

tagged_with(lexicon, tag)

The set of words or characters in a lexicon tagged with tag.

Coverage

CJKFrequencies.coverage — Function

coverage(lexicon, charfreq)
coverage(lexicon, text)
coverage([lexicon,] charfreq1, charfreq2)

Compute a "coverage" metric.

Between a lexicon of known words and a character frequency list, compute the coverage of the lexicon among characters in the character frequency list. Provide both token and type coverage:

token coverage counts each token separately (considering repeated characters)

type coverage counts each unique token once

Between a lexicon of known words and a string of text, compute the coverage of the lexicon among characters in the text. Also provide both token and type coverage.

Between 2 character frequency lists, compute the mutual coverages between the lists (optionally filtering using a provided lexicon). You can interpret these mutual coverage values as the (weighted) amount of overlap between the character frequencies (Jaccard similarity, or intersection-over-union). The lower the coverage, the higher the "switching cost" in terms of vocabulary.